Fixed color coding calendar up a little
